+++ service/rds/api.go
@@ -2065,6 +2065,9 @@ func (c *RDS) CreateDBInstanceRequest(input *CreateDBInstanceInput) (req *reques
 //     The network type is invalid for the DB instance. Valid nework type values
 //     are IPV4 and DUAL.
 //
+//   - ErrCodeCertificateNotFoundFault "CertificateNotFound"
+//     CertificateIdentifier doesn't refer to an existing certificate.
+//
 // See also, https://docs.aws.amazon.com/goto/WebAPI/rds-2014-10-31/CreateDBInstance
 func (c *RDS) CreateDBInstance(input *CreateDBInstanceInput) (*CreateDBInstanceOutput, error) {
 	req, out := c.CreateDBInstanceRequest(input)
@@ -5034,6 +5037,12 @@ func (c *RDS) DescribeCertificatesRequest(input *DescribeCertificatesInput) (req
 //
 // Lists the set of CA certificates provided by Amazon RDS for this Amazon Web
 // Services account.
+//
+// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+// in the Amazon Aurora User Guide.
 //
 // Returns awserr.Error for service API and SDK errors. Use runtime type assertions
 // with awserr.Error's Code and Message methods to get detailed information about
@@ -17045,6 +17054,12 @@ func (s *CancelExportTaskOutput) SetWarningMessage(v string) *CancelExportTaskOu
 }
 
 // A CA certificate for an Amazon Web Services account.
+//
+// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+// in the Amazon Aurora User Guide.
 type Certificate struct {
 	_ struct{} `type:"structure"`
 
@@ -17136,6 +17151,54 @@ func (s *Certificate) SetValidFrom(v time.Time) *Certificate {
 
 // SetValidTill sets the ValidTill field's value.
 func (s *Certificate) SetValidTill(v time.Time) *Certificate {
+	s.ValidTill = &v
+	return s
+}
+
+// Returns the details of the DB instance’s server certificate.
+//
+// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+// in the Amazon Aurora User Guide.
+type CertificateDetails struct {
+	_ struct{} `type:"structure"`
+
+	// The CA identifier of the CA certificate used for the DB instance's server
+	// certificate.
+	CAIdentifier *string `type:"string"`
+
+	// The expiration date of the DB instance’s server certificate.
+	ValidTill *time.Time `type:"timestamp"`
+}
+
+// String returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s CertificateDetails) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s CertificateDetails) GoString() string {
+	return s.String()
+}
+
+// SetCAIdentifier sets the CAIdentifier field's value.
+func (s *CertificateDetails) SetCAIdentifier(v string) *CertificateDetails {
+	s.CAIdentifier = &v
+	return s
+}
+
+// SetValidTill sets the ValidTill field's value.
+func (s *CertificateDetails) SetValidTill(v time.Time) *CertificateDetails {
 	s.ValidTill = &v
 	return s
 }
@@ -18903,6 +18966,15 @@ type CreateCustomDBEngineVersionOutput struct {
 	// The status of the DB engine version, either available or deprecated.
 	Status *string `type:"string"`
 
+	// A list of the supported CA certificate identifiers.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
+	SupportedCACertificateIdentifiers []*string `type:"list"`
+
 	// A list of the character sets supported by this engine for the CharacterSetName
 	// parameter of the CreateDBInstance operation.
 	SupportedCharacterSets []*CharacterSet `locationNameList:"CharacterSet" type:"list"`
@@ -18939,6 +19011,10 @@ type CreateCustomDBEngineVersionOutput struct {
 	// A value that indicates whether the engine version supports Babelfish for
 	// Aurora PostgreSQL.
 	SupportsBabelfish *bool `type:"boolean"`
+
+	// A value that indicates whether the engine version supports rotating the server
+	// certificate without rebooting the DB instance.
+	SupportsCertificateRotationWithoutRestart *bool `type:"boolean"`
 
 	// A value that indicates whether you can use Aurora global databases with a
 	// specific DB engine version.
@@ -19084,6 +19160,12 @@ func (s *CreateCustomDBEngineVersionOutput) SetStatus(v string) *CreateCustomDBE
 	return s
 }
 
+// SetSupportedCACertificateIdentifiers sets the SupportedCACertificateIdentifiers field's value.
+func (s *CreateCustomDBEngineVersionOutput) SetSupportedCACertificateIdentifiers(v []*string) *CreateCustomDBEngineVersionOutput {
+	s.SupportedCACertificateIdentifiers = v
+	return s
+}
+
 // SetSupportedCharacterSets sets the SupportedCharacterSets field's value.
 func (s *CreateCustomDBEngineVersionOutput) SetSupportedCharacterSets(v []*CharacterSet) *CreateCustomDBEngineVersionOutput {
 	s.SupportedCharacterSets = v
@@ -19117,6 +19199,12 @@ func (s *CreateCustomDBEngineVersionOutput) SetSupportedTimezones(v []*Timezone)
 // SetSupportsBabelfish sets the SupportsBabelfish field's value.
 func (s *CreateCustomDBEngineVersionOutput) SetSupportsBabelfish(v bool) *CreateCustomDBEngineVersionOutput {
 	s.SupportsBabelfish = &v
+	return s
+}
+
+// SetSupportsCertificateRotationWithoutRestart sets the SupportsCertificateRotationWithoutRestart field's value.
+func (s *CreateCustomDBEngineVersionOutput) SetSupportsCertificateRotationWithoutRestart(v bool) *CreateCustomDBEngineVersionOutput {
+	s.SupportsCertificateRotationWithoutRestart = &v
 	return s
 }
 
@@ -20935,6 +21023,18 @@ type CreateDBInstanceInput struct {
 	// in the Amazon RDS User Guide.
 	BackupTarget *string `type:"string"`
 
+	// Specifies the CA certificate identifier to use for the DB instance’s server
+	// certificate.
+	//
+	// This setting doesn't apply to RDS Custom.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
+	CACertificateIdentifier *string `type:"string"`
+
 	// For supported engines, this value indicates that the DB instance should be
 	// associated with the specified CharacterSet.
 	//
@@ -21877,6 +21977,12 @@ func (s *CreateDBInstanceInput) SetBackupRetentionPeriod(v int64) *CreateDBInsta
 // SetBackupTarget sets the BackupTarget field's value.
 func (s *CreateDBInstanceInput) SetBackupTarget(v string) *CreateDBInstanceInput {
 	s.BackupTarget = &v
+	return s
+}
+
+// SetCACertificateIdentifier sets the CACertificateIdentifier field's value.
+func (s *CreateDBInstanceInput) SetCACertificateIdentifier(v string) *CreateDBInstanceInput {
+	s.CACertificateIdentifier = &v
 	return s
 }
 
@@ -26049,6 +26155,15 @@ type DBEngineVersion struct {
 	// The status of the DB engine version, either available or deprecated.
 	Status *string `type:"string"`
 
+	// A list of the supported CA certificate identifiers.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
+	SupportedCACertificateIdentifiers []*string `type:"list"`
+
 	// A list of the character sets supported by this engine for the CharacterSetName
 	// parameter of the CreateDBInstance operation.
 	SupportedCharacterSets []*CharacterSet `locationNameList:"CharacterSet" type:"list"`
@@ -26085,6 +26200,10 @@ type DBEngineVersion struct {
 	// A value that indicates whether the engine version supports Babelfish for
 	// Aurora PostgreSQL.
 	SupportsBabelfish *bool `type:"boolean"`
+
+	// A value that indicates whether the engine version supports rotating the server
+	// certificate without rebooting the DB instance.
+	SupportsCertificateRotationWithoutRestart *bool `type:"boolean"`
 
 	// A value that indicates whether you can use Aurora global databases with a
 	// specific DB engine version.
@@ -26230,6 +26349,12 @@ func (s *DBEngineVersion) SetStatus(v string) *DBEngineVersion {
 	return s
 }
 
+// SetSupportedCACertificateIdentifiers sets the SupportedCACertificateIdentifiers field's value.
+func (s *DBEngineVersion) SetSupportedCACertificateIdentifiers(v []*string) *DBEngineVersion {
+	s.SupportedCACertificateIdentifiers = v
+	return s
+}
+
 // SetSupportedCharacterSets sets the SupportedCharacterSets field's value.
 func (s *DBEngineVersion) SetSupportedCharacterSets(v []*CharacterSet) *DBEngineVersion {
 	s.SupportedCharacterSets = v
@@ -26263,6 +26388,12 @@ func (s *DBEngineVersion) SetSupportedTimezones(v []*Timezone) *DBEngineVersion 
 // SetSupportsBabelfish sets the SupportsBabelfish field's value.
 func (s *DBEngineVersion) SetSupportsBabelfish(v bool) *DBEngineVersion {
 	s.SupportsBabelfish = &v
+	return s
+}
+
+// SetSupportsCertificateRotationWithoutRestart sets the SupportsCertificateRotationWithoutRestart field's value.
+func (s *DBEngineVersion) SetSupportsCertificateRotationWithoutRestart(v bool) *DBEngineVersion {
+	s.SupportsCertificateRotationWithoutRestart = &v
 	return s
 }
 
@@ -26368,7 +26499,16 @@ type DBInstance struct {
 	BackupTarget *string `type:"string"`
 
 	// The identifier of the CA certificate for this DB instance.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
 	CACertificateIdentifier *string `type:"string"`
+
+	// The details of the DB instance's server certificate.
+	CertificateDetails *CertificateDetails `type:"structure"`
 
 	// If present, specifies the name of the character set that this instance is
 	// associated with.
@@ -26847,6 +26987,12 @@ func (s *DBInstance) SetBackupTarget(v string) *DBInstance {
 // SetCACertificateIdentifier sets the CACertificateIdentifier field's value.
 func (s *DBInstance) SetCACertificateIdentifier(v string) *DBInstance {
 	s.CACertificateIdentifier = &v
+	return s
+}
+
+// SetCertificateDetails sets the CertificateDetails field's value.
+func (s *DBInstance) SetCertificateDetails(v *CertificateDetails) *DBInstance {
+	s.CertificateDetails = v
 	return s
 }
 
@@ -29307,6 +29453,15 @@ type DeleteCustomDBEngineVersionOutput struct {
 	// The status of the DB engine version, either available or deprecated.
 	Status *string `type:"string"`
 
+	// A list of the supported CA certificate identifiers.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
+	SupportedCACertificateIdentifiers []*string `type:"list"`
+
 	// A list of the character sets supported by this engine for the CharacterSetName
 	// parameter of the CreateDBInstance operation.
 	SupportedCharacterSets []*CharacterSet `locationNameList:"CharacterSet" type:"list"`
@@ -29343,6 +29498,10 @@ type DeleteCustomDBEngineVersionOutput struct {
 	// A value that indicates whether the engine version supports Babelfish for
 	// Aurora PostgreSQL.
 	SupportsBabelfish *bool `type:"boolean"`
+
+	// A value that indicates whether the engine version supports rotating the server
+	// certificate without rebooting the DB instance.
+	SupportsCertificateRotationWithoutRestart *bool `type:"boolean"`
 
 	// A value that indicates whether you can use Aurora global databases with a
 	// specific DB engine version.
@@ -29488,6 +29647,12 @@ func (s *DeleteCustomDBEngineVersionOutput) SetStatus(v string) *DeleteCustomDBE
 	return s
 }
 
+// SetSupportedCACertificateIdentifiers sets the SupportedCACertificateIdentifiers field's value.
+func (s *DeleteCustomDBEngineVersionOutput) SetSupportedCACertificateIdentifiers(v []*string) *DeleteCustomDBEngineVersionOutput {
+	s.SupportedCACertificateIdentifiers = v
+	return s
+}
+
 // SetSupportedCharacterSets sets the SupportedCharacterSets field's value.
 func (s *DeleteCustomDBEngineVersionOutput) SetSupportedCharacterSets(v []*CharacterSet) *DeleteCustomDBEngineVersionOutput {
 	s.SupportedCharacterSets = v
@@ -29521,6 +29686,12 @@ func (s *DeleteCustomDBEngineVersionOutput) SetSupportedTimezones(v []*Timezone)
 // SetSupportsBabelfish sets the SupportsBabelfish field's value.
 func (s *DeleteCustomDBEngineVersionOutput) SetSupportsBabelfish(v bool) *DeleteCustomDBEngineVersionOutput {
 	s.SupportsBabelfish = &v
+	return s
+}
+
+// SetSupportsCertificateRotationWithoutRestart sets the SupportsCertificateRotationWithoutRestart field's value.
+func (s *DeleteCustomDBEngineVersionOutput) SetSupportsCertificateRotationWithoutRestart(v bool) *DeleteCustomDBEngineVersionOutput {
+	s.SupportsCertificateRotationWithoutRestart = &v
 	return s
 }
 
@@ -38974,6 +39145,12 @@ type ModifyCertificatesOutput struct {
 	_ struct{} `type:"structure"`
 
 	// A CA certificate for an Amazon Web Services account.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
 	Certificate *Certificate `type:"structure"`
 }
 
@@ -39342,6 +39519,15 @@ type ModifyCustomDBEngineVersionOutput struct {
 	// The status of the DB engine version, either available or deprecated.
 	Status *string `type:"string"`
 
+	// A list of the supported CA certificate identifiers.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
+	SupportedCACertificateIdentifiers []*string `type:"list"`
+
 	// A list of the character sets supported by this engine for the CharacterSetName
 	// parameter of the CreateDBInstance operation.
 	SupportedCharacterSets []*CharacterSet `locationNameList:"CharacterSet" type:"list"`
@@ -39378,6 +39564,10 @@ type ModifyCustomDBEngineVersionOutput struct {
 	// A value that indicates whether the engine version supports Babelfish for
 	// Aurora PostgreSQL.
 	SupportsBabelfish *bool `type:"boolean"`
+
+	// A value that indicates whether the engine version supports rotating the server
+	// certificate without rebooting the DB instance.
+	SupportsCertificateRotationWithoutRestart *bool `type:"boolean"`
 
 	// A value that indicates whether you can use Aurora global databases with a
 	// specific DB engine version.
@@ -39523,6 +39713,12 @@ func (s *ModifyCustomDBEngineVersionOutput) SetStatus(v string) *ModifyCustomDBE
 	return s
 }
 
+// SetSupportedCACertificateIdentifiers sets the SupportedCACertificateIdentifiers field's value.
+func (s *ModifyCustomDBEngineVersionOutput) SetSupportedCACertificateIdentifiers(v []*string) *ModifyCustomDBEngineVersionOutput {
+	s.SupportedCACertificateIdentifiers = v
+	return s
+}
+
 // SetSupportedCharacterSets sets the SupportedCharacterSets field's value.
 func (s *ModifyCustomDBEngineVersionOutput) SetSupportedCharacterSets(v []*CharacterSet) *ModifyCustomDBEngineVersionOutput {
 	s.SupportedCharacterSets = v
@@ -39556,6 +39752,12 @@ func (s *ModifyCustomDBEngineVersionOutput) SetSupportedTimezones(v []*Timezone)
 // SetSupportsBabelfish sets the SupportsBabelfish field's value.
 func (s *ModifyCustomDBEngineVersionOutput) SetSupportsBabelfish(v bool) *ModifyCustomDBEngineVersionOutput {
 	s.SupportsBabelfish = &v
+	return s
+}
+
+// SetSupportsCertificateRotationWithoutRestart sets the SupportsCertificateRotationWithoutRestart field's value.
+func (s *ModifyCustomDBEngineVersionOutput) SetSupportsCertificateRotationWithoutRestart(v bool) *ModifyCustomDBEngineVersionOutput {
+	s.SupportsCertificateRotationWithoutRestart = &v
 	return s
 }
 
@@ -40191,7 +40393,7 @@ type ModifyDBClusterInput struct {
 	//
 	// Example: my-cluster2
 	//
-	// Valid for: Aurora DB clusters only
+	// Valid for: Aurora DB clusters and Multi-AZ DB clusters
 	NewDBClusterIdentifier *string `type:"string"`
 
 	// A value that indicates that the DB cluster should be associated with the
@@ -40956,9 +41158,16 @@ type ModifyDBInstanceInput struct {
 	//    is running PostgreSQL 9.3.5.
 	BackupRetentionPeriod *int64 `type:"integer"`
 
-	// Specifies the certificate to associate with the DB instance.
+	// Specifies the CA certificate identifier to use for the DB instance’s server
+	// certificate.
 	//
 	// This setting doesn't apply to RDS Custom.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
 	CACertificateIdentifier *string `type:"string"`
 
 	// A value that indicates whether the DB instance is restarted when you rotate
@@ -44727,6 +44936,12 @@ type PendingModifiedValues struct {
 	BackupRetentionPeriod *int64 `type:"integer"`
 
 	// The identifier of the CA certificate for the DB instance.
+	//
+	// For more information, see Using SSL/TLS to encrypt a connection to a DB instance
+	// (https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon RDS User Guide and Using SSL/TLS to encrypt a connection to
+	// a DB cluster (https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/UsingWithRDS.SSL.html)
+	// in the Amazon Aurora User Guide.
 	CACertificateIdentifier *string `type:"string"`
 
 	// The name of the compute and memory capacity class for the DB instance.
@@ -51717,26 +51932,26 @@ type StartExportTaskInput struct {
 	// The ID of the Amazon Web Services KMS key to use to encrypt the snapshot
 	// exported to Amazon S3. The Amazon Web Services KMS key identifier is the
 	// key ARN, key ID, alias ARN, or alias name for the KMS key. The caller of
-	// this operation must be authorized to execute the following operations. These
+	// this operation must be authorized to run the following operations. These
 	// can be set in the Amazon Web Services KMS key policy:
 	//
-	//    * GrantOperation.Encrypt
+	//    * kms:Encrypt
 	//
-	//    * GrantOperation.Decrypt
+	//    * kms:Decrypt
 	//
-	//    * GrantOperation.GenerateDataKey
+	//    * kms:GenerateDataKey
 	//
-	//    * GrantOperation.GenerateDataKeyWithoutPlaintext
+	//    * kms:GenerateDataKeyWithoutPlaintext
 	//
-	//    * GrantOperation.ReEncryptFrom
+	//    * kms:ReEncryptFrom
 	//
-	//    * GrantOperation.ReEncryptTo
+	//    * kms:ReEncryptTo
 	//
-	//    * GrantOperation.CreateGrant
+	//    * kms:CreateGrant
 	//
-	//    * GrantOperation.DescribeKey
+	//    * kms:DescribeKey
 	//
-	//    * GrantOperation.RetireGrant
+	//    * kms:RetireGrant
 	//
 	// KmsKeyId is a required field
 	KmsKeyId *string `type:"string" required:"true"`